new chunk stitcher done
The former script used to fill the unwanted gaps between two chunks used to generate a new mesh between the chunk edges to fill the gap. INEFFICIENT.
This update entirely changes the chunk stitching system, now I take the average of the two heights and make both of the height equal to that, then update both of the chunks. Works fine for now.